+++ packages/core/src/analyzers/javascript-analyzer.ts
@@ -3,7 +3,7 @@ import * as parser from '@babel/parser';
 import traverse, { TraverseOptions } from '@babel/traverse';
 import AstTraverser from '../ast/ast-traverser';
 
-export default class JavaScriptTraverser extends AstTraverser<
+export default class JavaScriptAnalyzer extends AstTraverser<
   File,
   TraverseOptions<Node>,
   typeof parser.parse,
