+++ packages/core/src/analyzers/json-analyzer.ts
@@ -1,6 +1,6 @@
-import JavaScriptTraverser from './javascript-traverser';
+import JavaScriptAnalyzer from './javascript-analyzer';
 
-export default class JsonTraverser extends JavaScriptTraverser {
+export default class JsonAnalyzer extends JavaScriptAnalyzer {
   constructor(source: string) {
     // In order to process JSON, we need to convert it to a module,
     // as babel cannot parse JSON natively.
