+++ packages/core/src/index.ts
@@ -27,9 +27,9 @@ export { default as TaskError } from './errors/task-error';
 export { default as AstTraverser } from './ast/ast-traverser';
 export { default as AstTransformer } from './ast/ast-transformer';
 
-export { default as JavaScriptTraverser } from './analyzers/javascript-traverser';
-export { default as TypeScriptTraverser } from './analyzers/typescript-traverser';
-export { default as JsonTraverser } from './analyzers/json-traverser';
+export { default as JavaScriptAnalyzer } from './analyzers/javascript-analyzer';
+export { default as TypeScriptAnalyzer } from './analyzers/typescript-analyzer';
+export { default as JsonAnalyzer } from './analyzers/json-analyzer';
 
 export { getPluginName, normalizePackageName, getShorthandName } from './utils/plugin-name';
 export { exec } from './utils/exec';
